refactor: move Copy component into ui folder
